# openapi-v2.yaml
openapi: 3.0.1
info:
  title: Contoso Real Estate API
  description: Finds up to date and detailed real estate properties for sale on the market
  version: 0.2.0
servers:
  - url: https://kev-contoso-real-estate-plugin-app.azurewebsites.net
paths:
  /get-dubai:
    get:
      operationId: getDubaiListings
      summary: Get listings in Dubai
      description: Returns a list of real estate listings with five bedrooms in Dubai.
      responses:
        '200':
          description: A list of listings in Dubai
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Listing'
        '400':
          description: Error message
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /get-listings:
    get:
      operationId: getListings
      summary: Get a list of properties matching the specified criteria
      description: Returns a list of properties, optionally filtered by providing the city, number of bedrooms, number of bathrooms and a list of amenities as query parameters
      parameters:
        - name: city
          in: query
          description: The city to search in
          required: false
          schema:
            type: string
        - name: bedrooms
          in: query
          description: The number of bedrooms
          required: false
          schema:
            type: integer
        - name: bathrooms
          in: query
          description: The number of bathrooms
          required: false
          schema:
            type: integer
        - name: amenities
          in: query
          description: The list of amenities
          required: false
          schema:
            $ref: '#/components/schemas/AmenitiesEnum'
      responses:
        '200':
          description: OK - Returns a list of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
  /post-a-booking:
    post:
      operationId: postBooking
      summary: Add a new booking
      description: Adds a new booking with the specified property ID and a date-time for the next day at 14:00 hrs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                property-id:
                  type: string
                  description: The ID of the property to book.
      responses:
        '200':
          description: Booking has been successfully added.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /bookings:
    get:
      operationId: getLastBooking
      summary: Get the last booking
      description: Returns the last booking added to the system.
      responses:
        '200':
          description: The last booking
          content:
            application/json:
              schema:
                type: object
                properties:
                  property-id:
                    type: string
                  date-time:
                    type: string
                    format: date-time
components:
  schemas:
    AmenitiesEnum:
      type: string
      enum:
        - air conditioning
        - balcony
        - dishwasher
        - elevator
        - fireplace
        - furniture
        - garden
        - gym
        - heating
        - jacuzzi
        - laundry room
        - microwave
        - no furniture
        - parking
        - patio
        - sauna
        - swimming pool
        - terrace
        - wi-fi
    Property:
      type: object
      properties:
        title:
          type: string
          description: The title of the property listing
        city:
          type: string
          description: The city of the property
        bedrooms:
          type: integer
          description: The number of bedrooms
        bathrooms:
          type: integer
          description: The number of bathrooms
        amenities:
          type: array
          description: The list of amenities
          items:
            $ref: '#/components/schemas/AmenitiesEnum'
        description:
          type: string
          description: The description of the property